Switched to the new structure without Kurogo.
Switched to the new structure without Kurogo.
The new authentication is going to be provided by mobilemanual and
mobileaai auth plugins. For every course a new token is going to be
provided. This way every request is going to have a restricted
context and the correct user.
Added a logging function. The log ends in the course releated log.