User based course modules info extraction
This commit fixes the course modules info extraction
based on the user. The Activities and resources that are invisible
for the user are not in the output anymore.

The visibility field is not part of the output structure.

This is the example of the new structure for the webservice
local_uniappws_get_course_modules:

[
	{
		"id": 11,
		"instanceid": 4,
		"courseid": 3,
		"modname": "forum",
		"name": "News forum",
		"intro": "General news and announcements",
		"timemodified": 1315833803
	},
	{
		"id": 272,
		"instanceid": 8,
		"courseid": 3,
		"modname": "forum",
		"name": "Nippon forum",
		"intro": "Nippon Omedeto gosaimasu",
		"timemodified": 1338891558
	}
]