Started Moodbile project integration.
Moodbile forum services added and adapted to Moodle 2.2.x version.
Reimplemented testing functions to use a real course. This way the
mock objects are not necessary and dealing with moodle contex
cache is not necessary. Obviously to perform the tests one must
have the test_course in place.
Added test_course backup with the testing forum.