Added Private files web service; Dropped tests
Added Private files web service. The code has been adapted to use
the standard pluginfile.php and not the custom one.

Dropped all simpletest folders containing tests. The moodle 2.3
testing suite is PHPUnit based and it is started from bash by
using phpunit command. The testing suite that was included in
this project was Unit based and dependant on a local testing course.
This way is possible to skip Mock objects entirely and to access
the DB. No Mock objects means no dependency on internal moodle
structure but it also means that the test is confined to a specific
installation. Once the installation changes the test suite has to be
configured to work on the new installation.

Locally, I'm still going to use my testing suite and once I'll have
time I'll write a new test suite PHPUnit compatible.